// src/app/components/shared/interfaces/item.interface.ts
import { Artist } from './artist.interface';
import { Image } from './image.interface';
import { ExternalUrls } from './external-urls.interface';

export interface Item {
  album_type: string;
  artists: Artist[];
  available_markets: string[];
  external_urls: ExternalUrls;
  href: string;
  id: string;
  images: Image[];
  name: string;
  release_date: string;
  release_date_precision: string;
  total_tracks: number;
  type: string;
  uri: string;
}